+++ Chartboost/src/main/java/com/mopub/mobileads/ChartboostRewardedVideo.java
@@ -32,6 +32,7 @@ public class ChartboostRewardedVideo extends CustomEventRewardedVideo {
     private final Handler mHandler;
 
     private static final String ADAPTER_NAME = ChartboostRewardedVideo.class.getSimpleName();
+    private static final String CUSTOM_ID_KEY = "customId";
 
     @NonNull
     private ChartboostAdapterConfiguration mChartboostAdapterConfiguration;
@@ -65,7 +66,7 @@ public class ChartboostRewardedVideo extends CustomEventRewardedVideo {
                                          @NonNull Map<String, String> serverExtras) throws Exception {
         // We need to attempt to reinitialize Chartboost on each request, in case an interstitial has been
         // loaded and used since then.
-        ChartboostShared.initializeSdk(launcherActivity, serverExtras);  // throws IllegalStateException
+        ChartboostShared.initializeSdk(launcherActivity.getApplicationContext(), serverExtras);  // throws IllegalStateException
 
         // Always return true so that the lifecycle listener is registered even if an interstitial
         // did the initialization.
@@ -85,7 +86,7 @@ public class ChartboostRewardedVideo extends CustomEventRewardedVideo {
         }
 
         ChartboostShared.getDelegate().registerRewardedVideoLocation(mLocation);
-        setUpMediationSettingsForRequest((String) localExtras.get(DataKeys.AD_UNIT_ID_KEY));
+        setUpMediationSettingsForRequest((String) localExtras.get(DataKeys.AD_UNIT_ID_KEY), localExtras);
 
         // We do this to ensure that the custom event manager has a chance to get the listener
         // and ad unit ID before any delegate callbacks are made.
@@ -101,14 +102,21 @@ public class ChartboostRewardedVideo extends CustomEventRewardedVideo {
         });
     }
 
-    private void setUpMediationSettingsForRequest(String moPubId) {
+    private void setUpMediationSettingsForRequest(String moPubId, Map<String, Object> localExtras) {
         final ChartboostMediationSettings globalSettings =
                 MoPubRewardedVideoManager.getGlobalMediationSettings(ChartboostMediationSettings.class);
         final ChartboostMediationSettings instanceSettings =
                 MoPubRewardedVideoManager.getInstanceMediationSettings(ChartboostMediationSettings.class, moPubId);
 
-        // Instance settings override global settings.
-        if (instanceSettings != null) {
+        final Object customIdObject = localExtras.get(CUSTOM_ID_KEY);
+
+        if (customIdObject instanceof String) {
+            String customId = (String) customIdObject;
+
+            if (!TextUtils.isEmpty(customId)) {
+                Chartboost.setCustomId(customId);
+            }
+        } else if (instanceSettings != null) {
             Chartboost.setCustomId(instanceSettings.getCustomId());
         } else if (globalSettings != null) {
             Chartboost.setCustomId(globalSettings.getCustomId());
@@ -122,7 +130,7 @@ public class ChartboostRewardedVideo extends CustomEventRewardedVideo {
 
     @Override
     public void showVideo() {
-        MoPubLog.log(SHOW_ATTEMPTED, ADAPTER_NAME);
+        MoPubLog.log(getAdNetworkId(), SHOW_ATTEMPTED, ADAPTER_NAME);
 
         if (hasVideoAvailable()) {
             Chartboost.showRewardedVideo(mLocation);
@@ -132,10 +140,10 @@ public class ChartboostRewardedVideo extends CustomEventRewardedVideo {
                     getAdNetworkId(),
                     MoPubErrorCode.NETWORK_NO_FILL);
 
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Attempted to show Chartboost rewarded video before it " +
-                    "was available.");
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Attempted to show Chartboost " +
+                    "rewarded video before it was available.");
 
-            MoPubLog.log(SHOW_FAILED, ADAPTER_NAME,
+            MoPubLog.log(getAdNetworkId(), SHOW_FAILED, ADAPTER_NAME,
                     MoPubErrorCode.NETWORK_NO_FILL.getIntCode(),
                     MoPubErrorCode.NETWORK_NO_FILL);
         }
@@ -150,22 +158,18 @@ public class ChartboostRewardedVideo extends CustomEventRewardedVideo {
     private static final class ChartboostLifecycleListener implements LifecycleListener {
         @Override
         public void onCreate(@NonNull Activity activity) {
-            Chartboost.onCreate(activity);
         }
 
         @Override
         public void onStart(@NonNull Activity activity) {
-            Chartboost.onStart(activity);
         }
 
         @Override
         public void onPause(@NonNull Activity activity) {
-            Chartboost.onPause(activity);
         }
 
         @Override
         public void onResume(@NonNull Activity activity) {
-            Chartboost.onResume(activity);
         }
 
         @Override
@@ -174,12 +178,10 @@ public class ChartboostRewardedVideo extends CustomEventRewardedVideo {
 
         @Override
         public void onStop(@NonNull Activity activity) {
-            Chartboost.onStop(activity);
         }
 
         @Override
         public void onDestroy(@NonNull Activity activity) {
-            Chartboost.onDestroy(activity);
         }
 
         @Override