+++ Vungle/src/main/java/com/mopub/mobileads/VungleInterstitial.java
@@ -3,6 +3,7 @@ package com.mopub.mobileads;
 import android.content.Context;
 import android.os.Handler;
 import android.os.Looper;
+
 import androidx.annotation.Keep;
 import androidx.annotation.NonNull;
 
@@ -27,14 +28,6 @@ import static com.mopub.common.logging.MoPubLog.AdapterLogEvent.SHOW_SUCCESS;
 public class VungleInterstitial extends CustomEventInterstitial {
 
     /*
-     * APP_ID_KEY is intended for MoPub internal use. Do not modify.
-     */
-    private static final String APP_ID_KEY = "appId";
-    private static final String PLACEMENT_ID_KEY = "pid";
-    private static final String PLACEMENT_IDS_KEY = "pids";
-    private static final String ADAPTER_NAME = VungleInterstitial.class.getSimpleName();
-
-    /*
      * These keys can be used with MoPubInterstitial.setLocalExtras()
      * to pass additional parameters to the SDK.
      */
@@ -47,6 +40,14 @@ public class VungleInterstitial extends CustomEventInterstitial {
     @Deprecated
     public static final String AD_ORIENTATION_KEY = "vungleAdOrientation";
 
+    /*
+     * APP_ID_KEY is intended for MoPub internal use. Do not modify.
+     */
+    private static final String APP_ID_KEY = "appId";
+    private static final String PLACEMENT_ID_KEY = "pid";
+    private static final String PLACEMENT_IDS_KEY = "pids";
+    private static final String ADAPTER_NAME = VungleInterstitial.class.getSimpleName();
+
     private static VungleRouter sVungleRouter;
     private final Handler mHandler;
     private CustomEventInterstitialListener mCustomEventInterstitialListener;
@@ -54,10 +55,9 @@ public class VungleInterstitial extends CustomEventInterstitial {
     @NonNull
     private VungleAdapterConfiguration mVungleAdapterConfiguration;
     private String mAppId;
-    private String mPlacementId;
+    private static String mPlacementId;
     private AdConfig mAdConfig;
     private boolean mIsPlaying;
-
 
     public VungleInterstitial() {
         mHandler = new Handler(Looper.getMainLooper());
@@ -81,67 +81,61 @@ public class VungleInterstitial extends CustomEventInterstitial {
                 public void run() {
                     mCustomEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.NETWORK_NO_FILL);
 
-                    MoPubLog.log(LOAD_FAILED, ADAPTER_NAME,
+                    MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
                             MoPubErrorCode.NETWORK_NO_FILL.getIntCode(),
                             MoPubErrorCode.NETWORK_NO_FILL);
                 }
             });
-
             return;
         }
-
         if (!validateIdsInServerExtras(serverExtras)) {
             mHandler.post(new Runnable() {
                 @Override
                 public void run() {
                     mCustomEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.NETWORK_NO_FILL);
 
-                    MoPubLog.log(LOAD_FAILED, ADAPTER_NAME,
+                    MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
                             MoPubErrorCode.NETWORK_NO_FILL.getIntCode(),
                             MoPubErrorCode.NETWORK_NO_FILL);
                 }
             });
-
             return;
         }
 
         if (mVungleRouterListener == null) {
             mVungleRouterListener = new VungleInterstitialRouterListener();
         }
-
         if (!sVungleRouter.isVungleInitialized()) {
             // No longer passing the placement IDs (pids) param per Vungle 6.3.17
             sVungleRouter.initVungle(context, mAppId);
             mVungleAdapterConfiguration.setCachedInitializationParameters(context, serverExtras);
         }
-
         if (localExtras != null) {
             mAdConfig = new AdConfig();
             VungleMediationConfiguration.adConfigWithLocalExtras(mAdConfig, localExtras);
         }
 
         sVungleRouter.loadAdForPlacement(mPlacementId, mVungleRouterListener);
-        MoPubLog.log(mPlacementId, LOAD_ATTEMPTED, ADAPTER_NAME);
+        MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
     }
 
     @Override
     protected void showInterstitial() {
-        MoPubLog.log(SHOW_ATTEMPTED, ADAPTER_NAME);
+        MoPubLog.log(getAdNetworkId(), SHOW_ATTEMPTED, ADAPTER_NAME);
 
         if (sVungleRouter.isAdPlayableForPlacement(mPlacementId)) {
-
             sVungleRouter.playAdForPlacement(mPlacementId, mAdConfig);
             mIsPlaying = true;
         } else {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "SDK tried to show a Vungle interstitial ad before it " +
-                    "finished loading. Please try again.");
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME,
+                    "SDK tried to show a Vungle interstitial ad before it finished loading. Please try again.");
             mHandler.post(new Runnable() {
 
                 @Override
                 public void run() {
                     mCustomEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.NETWORK_NO_FILL);
 
-                    MoPubLog.log(SHOW_FAILED, ADAPTER_NAME,
+                    MoPubLog.log(getAdNetworkId(), SHOW_FAILED, ADAPTER_NAME,
                             MoPubErrorCode.NETWORK_NO_FILL.getIntCode(),
                             MoPubErrorCode.NETWORK_NO_FILL);
                 }
@@ -151,7 +145,8 @@ public class VungleInterstitial extends CustomEventInterstitial {
 
     @Override
     protected void onInvalidate() {
-        MoPubLog.log(CUSTOM, ADAPTER_NAME, "onInvalidate is called for Placement ID:" + mPlacementId);
+        MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME,
+                "onInvalidate is called for Placement ID:" + mPlacementId);
         sVungleRouter.removeRouterListener(mPlacementId);
         mVungleRouterListener = null;
         mAdConfig = null;
@@ -164,27 +159,28 @@ public class VungleInterstitial extends CustomEventInterstitial {
         if (serverExtras.containsKey(APP_ID_KEY)) {
             mAppId = serverExtras.get(APP_ID_KEY);
             if (mAppId != null && mAppId.isEmpty()) {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "App ID is empty.");
+                MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "App ID is empty.");
                 isAllDataValid = false;
             }
         } else {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "AppID is not in serverExtras.");
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "AppID is not in serverExtras.");
             isAllDataValid = false;
         }
 
         if (serverExtras.containsKey(PLACEMENT_ID_KEY)) {
             mPlacementId = serverExtras.get(PLACEMENT_ID_KEY);
             if (mPlacementId != null && mPlacementId.isEmpty()) {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "Placement ID for this Ad Unit is empty.");
+                MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Placement ID for this Ad Unit is empty.");
                 isAllDataValid = false;
             }
         } else {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Placement ID for this Ad Unit is not in serverExtras.");
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME,
+                    "Placement ID for this Ad Unit is not in serverExtras.");
             isAllDataValid = false;
         }
 
         if (serverExtras.containsKey(PLACEMENT_IDS_KEY)) {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME + "No need to set placement IDs " +
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME + "No need to set placement IDs " +
                     "in MoPub dashboard with Vungle SDK version " +
                     com.vungle.warren.BuildConfig.VERSION_NAME);
         }
@@ -192,19 +188,29 @@ public class VungleInterstitial extends CustomEventInterstitial {
         return isAllDataValid;
     }
 
+    private static String getAdNetworkId() {
+        return mPlacementId;
+    }
+
     /*
      * VungleRouterListener
      */
     private class VungleInterstitialRouterListener implements VungleRouterListener {
-
         @Override
-        public void onAdEnd(String id) {
-            if (mPlacementId.equals(id)) {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdEnd - Placement ID: " + id);
+        public void onAdEnd(@NonNull String placementReferenceId, final boolean wasSuccessfulView,
+                            final boolean wasCallToActionClicked) {
+            if (mPlacementId.equals(placementReferenceId)) {
+                MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "onAdEnd - Placement ID: " +
+                        placementReferenceId + ", wasSuccessfulView: " + wasSuccessfulView +
+                        ", wasCallToActionClicked: " + wasCallToActionClicked);
                 mIsPlaying = false;
                 mHandler.post(new Runnable() {
                     @Override
                     public void run() {
+                        if (wasCallToActionClicked) {
+                            mCustomEventInterstitialListener.onInterstitialClicked();
+                            MoPubLog.log(getAdNetworkId(), CLICKED, ADAPTER_NAME);
+                        }
                         mCustomEventInterstitialListener.onInterstitialDismissed();
                     }
                 });
@@ -213,45 +219,17 @@ public class VungleInterstitial extends CustomEventInterstitial {
         }
 
         @Override
-        public void onAdClick(String id) {
-            if (mPlacementId.equals(id)) {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdClick - Placement ID: " + id);
-                mIsPlaying = false;
-                mHandler.post(new Runnable() {
-                    @Override
-                    public void run() {
-                        mCustomEventInterstitialListener.onInterstitialClicked();
-                        MoPubLog.log(CLICKED, ADAPTER_NAME);
-                    }
-                });
-            }
-        }
-
-        @Override
-        public void onAdRewarded(String id) {
-            //nothing to do
-        }
-
-        @Override
-        public void onAdLeftApplication(String id) {
-            //Nothing to do. If we call mCustomEventInterstitialListener.onLeaveApplication() it will cause
-            // onInterstitialClicked() event be called twice.
-        }
-
-        @Override
         public void onAdStart(@NonNull String placementReferenceId) {
             if (mPlacementId.equals(placementReferenceId)) {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdStart - Placement ID: " + placementReferenceId);
-
+                MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME,
+                        "onAdStart - Placement ID: " + placementReferenceId);
                 mIsPlaying = true;
-
                 mHandler.post(new Runnable() {
                     @Override
                     public void run() {
                         mCustomEventInterstitialListener.onInterstitialShown();
                         mCustomEventInterstitialListener.onInterstitialImpression();
-
-                        MoPubLog.log(SHOW_SUCCESS, ADAPTER_NAME);
+                        MoPubLog.log(getAdNetworkId(), SHOW_SUCCESS, ADAPTER_NAME);
                     }
                 });
             }
@@ -260,16 +238,14 @@ public class VungleInterstitial extends CustomEventInterstitial {
         @Override
         public void onUnableToPlayAd(@NonNull String placementReferenceId, String reason) {
             if (mPlacementId.equals(placementReferenceId)) {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "onUnableToPlayAd - Placement ID: " +
+                MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "onUnableToPlayAd - Placement ID: " +
                         placementReferenceId + ", reason: " + reason);
                 mIsPlaying = false;
-
                 mHandler.post(new Runnable() {
                     @Override
                     public void run() {
                         mCustomEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.NETWORK_NO_FILL);
-
-                        MoPubLog.log(LOAD_FAILED, ADAPTER_NAME,
+                        MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
                                 MoPubErrorCode.NETWORK_NO_FILL.getIntCode(),
                                 MoPubErrorCode.NETWORK_NO_FILL);
                     }
@@ -282,26 +258,23 @@ public class VungleInterstitial extends CustomEventInterstitial {
             if (mPlacementId.equals(placementReferenceId)) {
                 if (!mIsPlaying) {
                     if (isAdAvailable) {
-                        MoPubLog.log(CUSTOM, ADAPTER_NAME, "interstitial ad successfully loaded - " +
-                                "Placement ID: " + placementReferenceId);
-
+                        MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME,
+                                "interstitial ad successfully loaded - Placement ID: " + placementReferenceId);
                         mHandler.post(new Runnable() {
                             @Override
                             public void run() {
                                 mCustomEventInterstitialListener.onInterstitialLoaded();
-                                MoPubLog.log(LOAD_SUCCESS, ADAPTER_NAME);
+                                MoPubLog.log(getAdNetworkId(), LOAD_SUCCESS, ADAPTER_NAME);
                             }
                         });
                     } else {
-                        MoPubLog.log(CUSTOM, ADAPTER_NAME, "interstitial ad is not loaded - Placement ID: "
-                                + placementReferenceId);
-
+                        MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME,
+                                "interstitial ad is not loaded - Placement ID: " + placementReferenceId);
                         mHandler.post(new Runnable() {
                             @Override
                             public void run() {
                                 mCustomEventInterstitialListener.onInterstitialFailed(MoPubErrorCode.NETWORK_NO_FILL);
-
-                                MoPubLog.log(LOAD_FAILED, ADAPTER_NAME,
+                                MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
                                         MoPubErrorCode.NETWORK_NO_FILL.getIntCode(),
                                         MoPubErrorCode.NETWORK_NO_FILL);
                             }