+++ Vungle/src/main/java/com/mopub/mobileads/VungleRouter.java
@@ -2,15 +2,16 @@ package com.mopub.mobileads;
 
 import android.app.Activity;
 import android.content.Context;
+
 import androidx.annotation.NonNull;
 
 import com.mopub.common.BaseLifecycleListener;
 import com.mopub.common.LifecycleListener;
 import com.mopub.common.MoPub;
+import com.mopub.common.Preconditions;
 import com.mopub.common.logging.MoPubLog;
 import com.mopub.common.privacy.ConsentStatus;
 import com.mopub.common.privacy.PersonalInfoManager;
-
 import com.vungle.warren.AdConfig;
 import com.vungle.warren.AdConfig.AdSize;
 import com.vungle.warren.Banners;
@@ -35,18 +36,6 @@ public class VungleRouter {
 
     private static final String ADAPTER_NAME = VungleRouter.class.getSimpleName();
 
-    private static VungleRouter instance = new VungleRouter();
-
-    private enum SDKInitState {
-        NOTINITIALIZED,
-        INITIALIZING,
-        INITIALIZED
-    }
-
-    private static SDKInitState sInitState = SDKInitState.NOTINITIALIZED;
-    private static Map<String, VungleRouterListener> sVungleRouterListeners = new HashMap<>();
-    private static Map<String, VungleRouterListener> sWaitingList = new HashMap<>();
-
     private static final LifecycleListener sLifecycleListener = new BaseLifecycleListener() {
         @Override
         public void onPause(@NonNull final Activity activity) {
@@ -58,6 +47,16 @@ public class VungleRouter {
             super.onResume(activity);
         }
     };
+    private static VungleRouter sInstance = new VungleRouter();
+    private static SDKInitState sInitState = SDKInitState.NOTINITIALIZED;
+    private static Map<String, VungleRouterListener> sVungleRouterListeners = new HashMap<>();
+    private static Map<String, VungleRouterListener> sWaitingList = new HashMap<>();
+
+    private enum SDKInitState {
+        NOTINITIALIZED,
+        INITIALIZING,
+        INITIALIZED
+    }
 
     private VungleRouter() {
         Plugin.addWrapperInfo(VungleApiClient.WrapperFramework.mopub,
@@ -65,7 +64,7 @@ public class VungleRouter {
     }
 
     static VungleRouter getInstance() {
-        return instance;
+        return sInstance;
     }
 
     LifecycleListener getLifecycleListener() {
@@ -73,16 +72,13 @@ public class VungleRouter {
     }
 
     void initVungle(final Context context, final String vungleAppId) {
-
         // Pass the user consent from the MoPub SDK to Vungle as per GDPR
         // Pass consentMessageVersion per Vungle 6.3.17:
         // https://support.vungle.com/hc/en-us/articles/360002922871#GDPRRecommendedImplementationInstructions
-        //no-op
         InitCallback initCallback = new InitCallback() {
             @Override
             public void onSuccess() {
                 MoPubLog.log(CUSTOM, ADAPTER_NAME, "SDK is initialized successfully.");
-
                 sInitState = SDKInitState.INITIALIZED;
 
                 clearWaitingList();
@@ -97,25 +93,25 @@ public class VungleRouter {
                     if (shouldAllowLegitimateInterest) {
                         if (personalInfoManager.getPersonalInfoConsentStatus() == ConsentStatus.EXPLICIT_NO
                                 || personalInfoManager.getPersonalInfoConsentStatus() == ConsentStatus.DNT
-                                || personalInfoManager.getPersonalInfoConsentStatus() == ConsentStatus.POTENTIAL_WHITELIST) {
+                                || personalInfoManager.getPersonalInfoConsentStatus() ==
+                                ConsentStatus.POTENTIAL_WHITELIST) {
                             Vungle.updateConsentStatus(Vungle.Consent.OPTED_OUT, "");
                         } else {
                             Vungle.updateConsentStatus(Vungle.Consent.OPTED_IN, "");
                         }
                     } else {
                         // Pass consentMessageVersion per Vungle 6.3.17:
-                        // https://support.vungle.com/hc/en-us/articles/360002922871#GDPRRecommendedImplementationInstructions
+                        // https://support.vungle.com/hc/en-us/articles/360002922871#
+                        // GDPRRecommendedImplementationInstructions
                         Vungle.updateConsentStatus(canCollectPersonalInfo ? Vungle.Consent.OPTED_IN :
                                 Vungle.Consent.OPTED_OUT, "");
                     }
                 }
-
             }
 
             @Override
             public void onError(VungleException throwable) {
                 MoPubLog.log(CUSTOM_WITH_THROWABLE, "Initialization failed.", throwable);
-
                 sInitState = SDKInitState.NOTINITIALIZED;
             }
 
@@ -132,8 +128,7 @@ public class VungleRouter {
         sInitState = SDKInitState.INITIALIZING;
     }
 
-    void setIncentivizedFields(String userID, String title, String body,
-                               String keepWatching, String close) {
+    void setIncentivizedFields(String userID, String title, String body, String keepWatching, String close) {
         Vungle.setIncentivizedFields(userID, title, body, keepWatching, close);
     }
 
@@ -152,13 +147,12 @@ public class VungleRouter {
     void loadAdForPlacement(String placementId, VungleRouterListener routerListener) {
         switch (sInitState) {
             case NOTINITIALIZED:
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "loadAdForPlacement is called before the adapter initialization.");
+                MoPubLog.log(placementId, CUSTOM, ADAPTER_NAME, "loadAdForPlacement is called before " +
+                        "initialization starts. This is not an expect case.");
                 break;
-
             case INITIALIZING:
                 sWaitingList.put(placementId, routerListener);
                 break;
-
             case INITIALIZED:
                 if (isValidPlacement(placementId)) {
                     addRouterListener(placementId, routerListener);
@@ -170,22 +164,23 @@ public class VungleRouter {
         }
     }
 
-    void loadBannerAd(@NonNull String placementId, @NonNull AdSize adSize, @NonNull VungleRouterListener routerListener) {
+    void loadBannerAd(@NonNull String placementId, @NonNull AdSize adSize,
+                      @NonNull VungleRouterListener routerListener) {
         switch (sInitState) {
             case NOTINITIALIZED:
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "loadBannerAdForPlacement is called before the adapter initialization.");
+                MoPubLog.log(CUSTOM, ADAPTER_NAME, "loadBannerAdForPlacement is called before the " +
+                        "Vungle SDK initialization.");
                 break;
-
             case INITIALIZING:
                 sWaitingList.put(placementId, routerListener);
                 break;
-
             case INITIALIZED:
                 if (isValidPlacement(placementId)) {
                     addRouterListener(placementId, routerListener);
                     Banners.loadBanner(placementId, adSize, loadAdCallback);
                 } else {
                     routerListener.onUnableToPlayAd(placementId, "Invalid/Inactive Banner Placement Id");
+                    MoPubLog.log(CUSTOM, ADAPTER_NAME, "Unable to play ad due to invalid/inactive Banner placement Id");
                 }
                 break;
         }
@@ -211,6 +206,9 @@ public class VungleRouter {
     }
 
     boolean isBannerAdPlayable(@NonNull String placementId, @NonNull AdSize adSize) {
+        Preconditions.checkNotNull(placementId);
+        Preconditions.checkNotNull(adSize);
+
         return Banners.canPlayAd(placementId, adSize);
     }
 
@@ -218,8 +216,8 @@ public class VungleRouter {
         if (isAdPlayableForPlacement(placementId)) {
             Vungle.playAd(placementId, adConfig, playAdCallback);
         } else {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "There should not be this case. playAdForPlacement is called " +
-                    "before an ad is loaded for Placement ID: " + placementId);
+            MoPubLog.log(placementId, CUSTOM, ADAPTER_NAME, "There should not be this case. " +
+                    "playAdForPlacement is called before an ad is loaded for Placement ID: " + placementId);
         }
     }
 
@@ -228,18 +226,20 @@ public class VungleRouter {
     }
 
     VungleBanner getVungleBannerAd(@NonNull String placementId, @NonNull AdSize adSize) {
+        Preconditions.checkNotNull(placementId);
+        Preconditions.checkNotNull(adSize);
+
         return Banners.getBanner(placementId, adSize, playAdCallback);
     }
 
     /**
      * Checks and returns if the passed Placement ID is a valid placement for App ID
      *
-     * @param placementId
+     * @param placementId placementReferenceId
      * @return
      */
     boolean isValidPlacement(String placementId) {
-        return Vungle.isInitialized() &&
-                Vungle.getValidPlacements().contains(placementId);
+        return Vungle.isInitialized() && Vungle.getValidPlacements().contains(placementId);
     }
 
     public void updateConsentStatus(Vungle.Consent status) {
@@ -264,15 +264,10 @@ public class VungleRouter {
     private final PlayAdCallback playAdCallback = new PlayAdCallback() {
         @Override
         public void onAdEnd(String id, boolean completed, boolean isCTAClicked) {
-            //Deprecated event
-        }
-
-        @Override
-        public void onAdEnd(String id) {
             MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdEnd - Placement ID: " + id);
             VungleRouterListener targetListener = sVungleRouterListeners.get(id);
             if (targetListener != null) {
-                targetListener.onAdEnd(id);
+                targetListener.onAdEnd(id, completed, isCTAClicked);
             } else {
                 MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdEnd - VungleRouterListener is not found for " +
                         "Placement ID: " + id);
@@ -280,50 +275,13 @@ public class VungleRouter {
         }
 
         @Override
-        public void onAdClick(String id) {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdClick - Placement ID: " + id);
-            VungleRouterListener targetListener = sVungleRouterListeners.get(id);
-            if (targetListener != null) {
-                targetListener.onAdClick(id);
-            } else {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdClick - VungleRouterListener is not found for " +
-                        "Placement ID: " + id);
-            }
-        }
-
-        @Override
-        public void onAdRewarded(String id) {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdRewarded - Placement ID: " + id);
-            VungleRouterListener targetListener = sVungleRouterListeners.get(id);
-            if (targetListener != null) {
-                targetListener.onAdRewarded(id);
-            } else {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdRewarded - VungleRouterListener is not found for " +
-                        "Placement ID: " + id);
-            }
-        }
-
-        @Override
-        public void onAdLeftApplication(String id) {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdLeftApplication - Placement ID: " + id);
-            VungleRouterListener targetListener = sVungleRouterListeners.get(id);
-            if (targetListener != null) {
-                targetListener.onAdLeftApplication(id);
-            } else {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdLeftApplication - VungleRouterListener is not found for " +
-                        "Placement ID: " + id);
-            }
-        }
-
-        @Override
         public void onAdStart(String id) {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdStart - Placement ID: " + id);
-
+            MoPubLog.log(id, CUSTOM, ADAPTER_NAME, "onAdStart - Placement ID: " + id);
             VungleRouterListener targetListener = sVungleRouterListeners.get(id);
             if (targetListener != null) {
                 targetListener.onAdStart(id);
             } else {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdStart - VungleRouterListener is not found for " +
+                MoPubLog.log(id, CUSTOM, ADAPTER_NAME, "onAdStart - VungleRouterListener is not found for " +
                         "Placement ID: " + id);
             }
         }
@@ -331,12 +289,11 @@ public class VungleRouter {
         @Override
         public void onError(String id, VungleException error) {
             MoPubLog.log(CUSTOM_WITH_THROWABLE, "onUnableToPlayAd - Placement ID: " + id, error);
-
             VungleRouterListener targetListener = sVungleRouterListeners.get(id);
             if (targetListener != null) {
                 targetListener.onUnableToPlayAd(id, error.getLocalizedMessage());
             } else {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "onUnableToPlayAd - VungleRouterListener is not found " +
+                MoPubLog.log(id, CUSTOM, ADAPTER_NAME, "onUnableToPlayAd - VungleRouterListener is not found " +
                         "for Placement ID: " + id);
             }
         }
@@ -354,15 +311,14 @@ public class VungleRouter {
         }
 
         private void onAdAvailabilityUpdate(String placementReferenceId, boolean isAdAvailable) {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdAvailabilityUpdate - Placement ID: " +
+            MoPubLog.log(placementReferenceId, CUSTOM, ADAPTER_NAME, "onAdAvailabilityUpdate - Placement ID: " +
                     placementReferenceId);
-
             VungleRouterListener targetListener = sVungleRouterListeners.get(placementReferenceId);
             if (targetListener != null) {
                 targetListener.onAdAvailabilityUpdate(placementReferenceId, isAdAvailable);
             } else {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "onAdAvailabilityUpdate - VungleRouterListener is not " +
-                        "found for Placement ID: " + placementReferenceId);
+                MoPubLog.log(placementReferenceId, CUSTOM, ADAPTER_NAME, "onAdAvailabilityUpdate - " +
+                        "VungleRouterListener is not found for Placement ID: " + placementReferenceId);
             }
         }
     };