+++ Chartboost/CHANGELOG.md
@@ -1,4 +1,15 @@
 ## Changelog
+  * 8.0.1.0
+    * This version of the adapters has been certified with Chartboost 8.0.1 and MoPub 5.11.1.
+    * Remove `Activity` requirement for Chartboost initialization, and remove `<activity>` declarations from the adapter AndroidManifest.
+
+  * 7.5.0.3
+    * Log the Chartboost location String in ad lifecycle events.
+    * Map additional error codes for failure cases.
+
+  * 7.5.0.2
+    * Prioritize reading data from `localExtras` and fall back to `MediationSettings` if necessary for ChartboostRewardedVideo.java.
+
   * 7.5.0.1
     * Add support for AndroidX. This is the minimum version compatible with MoPub 5.9.0.